Fix execution on JRE 11
Jotlmsg has a dependency on javax.activation because OutlookMessage makes use of it (for example, the ByteArrayDataSource). Java 9 removed the javax.activation package from the classpath. As a result, loading the OutlookMessage class on Java 11 causes a NoClassDefFoundError.

When an application declares a dependency on the optional com.sun.mail:javax.mail dependency, then the error does not manifest itself because com.sun.mail:javax.mail pulls in javax.activation:activation transitively.

On Java 8, everything works as expected, regardless of whether or not the optional dependency is specified, since javax.activation is on the classpath.